+++ ubbprotool-pc/src/main/java/com/chk/ubbprotool/ubbprotool/Service/MessageServiceImpl.java
@@ -33,7 +33,7 @@ public class MessageServiceImpl implements MessageService{
 
     @Override
     @Transactional
-    public void deleteMessage(int messageId) {
+    public void deleteMessage(Long messageId) {
         messageRepository.removeData(messageId);
     }
 
@@ -60,7 +60,7 @@ public class MessageServiceImpl implements MessageService{
 
     @Override
     @Transactional
-    public MessageDTO findById(int id) {
+    public MessageDTO findById(Long id) {
         Message message = messageRepository.findById(id);
         return messageMapper.toDTO(message);
     }
